feat: create TickerService, make DeribitClient async context manager

+++ app/models/ticker.py
@@ -9,4 +9,4 @@ class Ticker(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     name: Mapped[str] = mapped_column(String(10))
     price: Mapped[float] = mapped_column()
-    timestamp: Mapped[float] = mapped_column()
+    timestamp: Mapped[int] = mapped_column()
